ref: moved client class to client.py
