+++ clients/windows/mooshak/assets_loader.py
